issued(mass): Mass Transfer Token
adin and mass transfer
it is now possible:
  - with just one transaction to send 100 transfers for the price of 51.
the mass transfer price is given by: f(x) = 0.001 + 0.0005x
the price of a common transfer is given by g(x) = 0.001x
comparing:
f(x) > g(x) for x > 2
i.e. for more than 2 transfers of any asset,
it is cheaper to send it as a mass transfer.